requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
stripe==10.12.0
python-multipart==0.0.12
requests==2.32.3
apscheduler==3.10.4
